Change package to iam.snippets
