+++ iam/api-client/src/main/java/iam/snippets/DisableServiceAccount.java
@@ -13,57 +13,65 @@
  * limitations under the License.
  */
 
-package com.google.iam.snippets;
+package iam.snippets;
 
-// [START iam_get_policy]
+// [START iam_disable_service_account]
 import com.google.api.client.googleapis.auth.oauth2.GoogleCredential;
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport;
 import com.google.api.client.json.jackson2.JacksonFactory;
-import com.google.api.services.cloudresourcemanager.CloudResourceManager;
-import com.google.api.services.cloudresourcemanager.model.GetIamPolicyRequest;
-import com.google.api.services.cloudresourcemanager.model.Policy;
+import com.google.api.services.iam.v1.Iam;
 import com.google.api.services.iam.v1.IamScopes;
+import com.google.api.services.iam.v1.model.DisableServiceAccountRequest;
 import java.io.IOException;
 import java.security.GeneralSecurityException;
 import java.util.Collections;
 
-public class GetPolicy {
+public class DisableServiceAccount {
 
-  // Gets a project's policy.
-  public static Policy getPolicy(String projectId) {
-    // projectId = "my-project-id"
+  // Disables a service account.
+  public static void disableServiceAccount(String projectId) {
+    // String projectId = "my-project-id";
 
-    Policy policy = null;
-
-    CloudResourceManager service = null;
+    Iam service = null;
     try {
-      service = createCloudResourceManagerService();
+      service = initService();
     } catch (IOException | GeneralSecurityException e) {
       System.out.println("Unable to initialize service: \n" + e.toString());
-      return policy;
+      return;
     }
 
     try {
-      GetIamPolicyRequest request = new GetIamPolicyRequest();
-      policy = service.projects().getIamPolicy(projectId, request).execute();
-      System.out.println("Policy retrieved: " + policy.toString());
-      return policy;
+      DisableServiceAccountRequest request = new DisableServiceAccountRequest();
+      service
+          .projects()
+          .serviceAccounts()
+          .disable(
+              "projects/-/serviceAccounts/"
+                  + "your-service-account-name@"
+                  + projectId
+                  + ".iam.gserviceaccount.com",
+              request)
+          .execute();
+
+      System.out.println(
+          "Disabled service account: "
+              + "your-service-account-name@"
+              + projectId
+              + ".iam.gserviceaccount.com");
     } catch (IOException e) {
-      System.out.println("Unable to get policy: \n" + e.toString());
-      return policy;
+      System.out.println("Unable to disable service account: \n" + e.toString());
     }
   }
 
-  public static CloudResourceManager createCloudResourceManagerService()
-      throws IOException, GeneralSecurityException {
+  private static Iam initService() throws GeneralSecurityException, IOException {
     // Use the Application Default Credentials strategy for authentication. For more info, see:
     // https://cloud.google.com/docs/authentication/production#finding_credentials_automatically
     GoogleCredential credential =
         GoogleCredential.getApplicationDefault()
             .createScoped(Collections.singleton(IamScopes.CLOUD_PLATFORM));
-
-    CloudResourceManager service =
-        new CloudResourceManager.Builder(
+    // Initialize the IAM service, which can be used to send requests to the IAM API.
+    Iam service =
+        new Iam.Builder(
                 GoogleNetHttpTransport.newTrustedTransport(),
                 JacksonFactory.getDefaultInstance(),
                 credential)
@@ -72,4 +80,4 @@ public class GetPolicy {
     return service;
   }
 }
-// [END iam_get_policy]
+// [END iam_disable_service_account]
