+++ iam/api-client/src/main/java/iam/snippets/RemoveMember.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package com.google.iam.snippets;
+package iam.snippets;
 
 // [START iam_modify_policy_remove_member]
 import com.google.api.services.cloudresourcemanager.model.Binding;
