+++ iam/api-client/src/test/java/iam/snippets/ServiceAccountTests.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package com.google.iam.snippets;
+package iam.snippets;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
